Back end API Games final stage

+++ backend/src/API/Games/Delete.ts
@@ -3,19 +3,44 @@ import { GamesModel } from '../../models/GamesModel'
 
 export const Delete = async (req: Request, res: Response): Promise<void> => 
 {
-    const game_id: number = req.body.game_id;
+    // Permissions required: Creator of the game or administrator
+
+    if (!req.isAuthenticated()) {
+        res.json({
+            "status": "fail",
+            "error_message": "You need to be authenticated to do this operation",
+        });
+    }
+
+    const game_id: string = req.body.game_id;
 
     const game = await GamesModel.findById(game_id);
 
-    try {
-        game?.delete()
-        res.json({
-            "status": "ok"
-        });
-    } catch (error) {
+    if (game) {
+
+        if (!req.user || (req.user.IsAdministrator != true && req.user.id != game.AuthorID)) {
+            res.json({
+                "status": "fail",
+                "error_message": "Permission denied",
+            });
+        }
+
+        try {
+            game?.delete()
+            res.json({
+                "status": "ok",
+            });
+        } catch (error) {
+            res.json({
+                "status": "fail",
+                "error_message": error,
+            });
+        }
+    }
+    else {
         res.json({
             "status": "fail",
-            "error_message": error
+            "error_message": "Game Id not found",
         });
     }
